Update to new libraries

+++ cars/src/main/java/dk/aau/cs/giraf/cars/MainActivity.java
@@ -7,18 +7,18 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import dk.aau.cs.giraf.cars.Game.CarGame;
+import dk.aau.cs.giraf.cars.game.CarGame;
 import dk.aau.cs.giraf.gui.GButtonProfileSelect;
 import dk.aau.cs.giraf.gui.GComponent;
-import dk.aau.cs.giraf.oasis.lib.Helper;
-import dk.aau.cs.giraf.oasis.lib.models.Profile;
+import dk.aau.cs.giraf.dblib.Helper;
+import dk.aau.cs.giraf.dblib.models.Profile;
 
 public class MainActivity extends Activity {
 
     private static final int SETTINGS_IDENTIFIER = 0;
     private static final int MAPEDITOR_IDENTIFIER = 1;
 
-    int currentId, guardianId;
+    long currentId, guardianId;
     GButtonProfileSelect gButtonProfileSelect;
 
     @Override
@@ -29,7 +29,7 @@ public class MainActivity extends Activity {
 
         DatabaseHelper database = new DatabaseHelper(this);
 
-        //dk.aau.cs.giraf.oasis.lib.Helper h = new dk.aau.cs.giraf.oasis.lib.Helper(this);
+        //dk.aau.cs.giraf.dblib.Helper h = new dk.aau.cs.giraf.dblib.Helper(this);
         //h.CreateDummyData();
         // Skip loading screen if monkey test
         if (ActivityManager.isUserAMonkey()) {
@@ -40,9 +40,9 @@ public class MainActivity extends Activity {
             guardianId = h.profilesHelper.getGuardians().get(0).getId();
         }
         else {
-            currentId = intent.getIntExtra(DatabaseHelper.CHILD_ID, database.GetDefaultChild());
-            Log.d("id", Integer.toString(currentId));
-            guardianId = intent.getIntExtra(DatabaseHelper.GUARDIAN_ID, database.GetChildDefaultGuardian());
+            currentId = intent.getLongExtra(DatabaseHelper.CHILD_ID, database.GetDefaultChild());
+            Log.d("id", Long.toString(currentId));
+            guardianId = intent.getLongExtra(DatabaseHelper.GUARDIAN_ID, database.GetChildDefaultGuardian());
         }
 
         Profile curGuardian = database.GetProfileById(guardianId);
